app/tourguide-frontend-fresh/project structure/workspace/backend/server/requirements.txt
django==6.0.3
djangorestframework==3.16.1
django-cors-headers==4.9.0
channels==4.3.2
channels-redis==4.3.0
openai==1.68.2
python-dotenv==1.0.1
stripe==12.0.0